pyadmin/requirements.txt
absl-py==2.3.1
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
cffi==2.0.0
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
dnspython==2.8.0
ecdsa==0.19.1
fastapi==0.121.0
flatbuffers==25.2.10
fonttools==4.59.1
graphviz==0.21
h11==0.16.0
idna==3.11
jax==0.7.1
jaxlib==0.7.1
joblib==1.5.1
kiwisolver==1.4.9
matplotlib==3.10.5
mediapipe==0.10.14
ml_dtypes==0.5.3
motor==3.7.1
numpy==2.3.2
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opt_einsum==3.4.0
packaging==25.0
pillow==11.3.0
protobuf==4.25.8
pyasn1==0.6.1
pycparser==2.22
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
pymongo==4.15.3
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
rsa==4.9.1
scikit-learn==1.7.1
scipy==1.16.1
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
starlette==0.49.3
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
